@@ -11,7 +11,7 @@
     "migrate": "npx prisma migrate dev --skip-generate",
     "seed": "npx prisma db seed",
     "prebuild": "rimraf dist",
-    "prepare": "husky install",
+    "prepare": "husky",
     "start": "nest start",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
@@ -40,58 +40,58 @@
     "docker-prod-logs": "docker compose -f docker-compose.prod.yml logs"
   },
   "dependencies": {
-    "@nestjs/axios": "^3.0.3",
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/config": "^2.2.0",
-    "@nestjs/core": "^9.0.0",
-    "@nestjs/platform-express": "^9.0.0",
-    "@nestjs/swagger": "^6.1.4",
-    "@nestjs/throttler": "^4.0.0",
-    "@prisma/client": "^5.19.1",
+    "@nestjs/axios": "^4.0.0",
+    "@nestjs/common": "^11.0.7",
+    "@nestjs/config": "^4.0.0",
+    "@nestjs/core": "^11.0.7",
+    "@nestjs/platform-express": "^11.0.7",
+    "@nestjs/swagger": "^11.0.3",
+    "@nestjs/throttler": "^6.4.0",
+    "@prisma/client": "^6.3.1",
     "@types/bcryptjs": "^2.4.6",
-    "@types/cookie-parser": "^1.4.3",
-    "axios": "^1.7.7",
+    "@types/cookie-parser": "^1.4.8",
+    "axios": "^1.7.9",
     "bcryptjs": "^2.4.3",
     "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
-    "cookie-parser": "^1.4.6",
-    "express": "^4.18.2",
-    "joi": "^17.7.0",
-    "jsonwebtoken": "^9.0.0",
-    "mysql2": "^3.0.1",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0",
+    "class-validator": "^0.14.1",
+    "cookie-parser": "^1.4.7",
+    "express": "^4.21.2",
+    "joi": "^17.13.3",
+    "jsonwebtoken": "^9.0.2",
+    "mysql2": "^3.12.0",
+    "reflect-metadata": "^0.2.2",
+    "rimraf": "^6.0.1",
+    "rxjs": "^7.8.1",
     "swagger-themes": "^1.4.3",
-    "swagger-ui-express": "^4.6.0",
-    "uuid": "^9.0.0",
-    "webpack": "^5.77.0"
+    "swagger-ui-express": "^5.0.1",
+    "uuid": "^11.0.5",
+    "webpack": "^5.97.1"
   },
   "devDependencies": {
-    "@nestjs/cli": "^9.0.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.0.0",
-    "@types/express": "^4.17.13",
-    "@types/jest": "28.1.8",
-    "@types/node": "^16.0.0",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^8.0.0",
-    "jest": "28.1.3",
-    "lint-staged": "^13.2.0",
-    "prettier": "^2.3.2",
-    "prisma": "^5.19.1",
+    "@nestjs/cli": "^11.0.2",
+    "@nestjs/schematics": "^11.0.0",
+    "@nestjs/testing": "^11.0.7",
+    "@types/express": "^5.0.0",
+    "@types/jest": "29.5.14",
+    "@types/node": "^22.13.1",
+    "@types/supertest": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^8.23.0",
+    "@typescript-eslint/parser": "^8.23.0",
+    "eslint": "^9.19.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "husky": "^9.1.7",
+    "jest": "29.7.0",
+    "lint-staged": "^15.4.3",
+    "prettier": "^3.4.2",
+    "prisma": "^6.3.1",
     "source-map-support": "^0.5.20",
-    "supertest": "^6.1.3",
-    "ts-jest": "28.0.8",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "4.1.0",
-    "typescript": "^4.7.4"
+    "supertest": "^7.0.0",
+    "ts-jest": "29.2.5",
+    "ts-loader": "^9.5.2",
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "4.2.0",
+    "typescript": "^5.7.3"
   },
   "jest": {
     "moduleFileExtensions": [
